requirements.txt
watchdog==3.0.0
Pillow==9.4.0
transformers==4.34.0
googletrans==4.0.0-rc1
gTTS==2.3.2
pygame==2.6.1
